Clean up initialization and wasm wrapper
